WIP: Refactor docker client
